src/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "baykeshop"
version = "1.0.4"
authors = [
  { name="轻编程", email="1158920674@qq.com" },
]
keywords = ["django", "baykeshop"]
description = "python栈django框架，全开源商城系统baykeshop！"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "django",
  "alipay-sdk-python",
  "bayketinymce-django",
  "django-simple-captcha",
  "djangorestframework",
  "Pillow",
]

[project.urls]
"Homepage" = "http://www.bayke.shop"
"Bug Tracker" = "http://www.bayke.shop"